Portföy sayfası için gerekli Api bağlantıları yapıldı.

+++ GoldMarketCap_Kotlin/app/src/main/java/com/mulosbron/goldmarketcap/model/Portfolio.kt
@@ -4,8 +4,8 @@ data class Transaction(
     val id: String,
     val date: String,
     val transactionType: String,
-    val amount: Int,
-    val price: Int
+    val amount: Double,
+    val price: Double
 )
 
 data class TransactionResponse(
